Refactor, fix warnings
