+++ src/ast/astnode.h
@@ -45,7 +45,6 @@ enum class ASTType {
 };
 
 extern std::unordered_map<ASTType, std::string> ast_type_names;
-
 // operator precedence for each token
 extern std::unordered_map<ASTType, int> op_precedence;
 
@@ -58,90 +57,32 @@ class ASTNode {
   int _rbp = 0;
 
   ASTNode() = default;
-
-  ASTNode(ASTType op, int associativity, int lbp,
-          int rbp) : _op(op), _associativity(associativity), _lbp(lbp), _rbp(rbp) {};
-
-  virtual ~ASTNode() {
-      for (auto *&c : _children) {
-          delete c;
-          c = nullptr;
-      }
-  };
-
-  [[nodiscard]] virtual int get_ivalue() const {
-      assert(false);
-      return 0;
-  }
-
-  [[nodiscard]] virtual float get_fvalue() const {
-      assert(false);
-      return 0;
-  }
-
-  [[nodiscard]] virtual std::string get_svalue() const {
-      assert(false);
-      return "";
-  }
-
-  [[nodiscard]] virtual ASTNode *led(ASTNode *left, Parser *parser) {
-      assert(false);
-      return nullptr;
-  }
-
-  [[nodiscard]] virtual ASTNode *nud(Parser *parser) {
-      assert(false);
-      return nullptr;
-  }
-
-  virtual void add(ASTNode *c) {
-      _children.emplace_back(c);
-  }
-
+  ASTNode(ASTType op, int associativity, int lbp, int rbp);
+  virtual ~ASTNode();
+  [[nodiscard]] virtual int get_ivalue() const;
+  [[nodiscard]] virtual float get_fvalue() const;
+  [[nodiscard]] virtual std::string get_svalue() const;
+  [[nodiscard]] virtual ASTNode *led(ASTNode *left, Parser *parser);
+  [[nodiscard]] virtual ASTNode *nud(Parser *parser);
+  virtual void add(ASTNode *c);
   void printTree();
-
   void printSubtree(const std::string &prefix);
 };
 
 class ASTInfixBinaryOp : public ASTNode {
  public:
-  ASTInfixBinaryOp() : ASTNode(ASTType::EOF_, 0, op_precedence[ASTType::EOF_], 0) {
-  };
-
+  ASTInfixBinaryOp();
   [[nodiscard]] ASTNode *led(ASTNode *left, Parser *parser) override;
-
   [[nodiscard]] ASTNode *nud(Parser *parser) override;
 };
 
 class ASTNumberLiteral final : public ASTNode {
  public:
-  ASTNumberLiteral(const std::string &str, bool is_float) : ASTNode(ASTType::NUM_LITERAL, 1,
-                                                                    op_precedence[ASTType::NUM_LITERAL], 0) {
-      _is_float = is_float;
-      if (is_float) {
-          _fvalue = std::stof(str);
-      } else {
-          _ivalue = std::stoi(str);
-      }
-  };
-
-  [[nodiscard]] ASTNode *nud(Parser *parser) override {
-      return this;
-  }
-
-  [[nodiscard]] bool is_float() const {
-      return _is_float;
-  }
-
-  [[nodiscard]] int get_ivalue() const override {
-      assert(!_is_float);
-      return _ivalue;
-  }
-
-  [[nodiscard]] float get_fvalue() const override {
-      assert(_is_float);
-      return _fvalue;
-  }
+  ASTNumberLiteral(const std::string &str, bool is_float);
+  [[nodiscard]] ASTNode *nud(Parser *parser) override;
+  [[nodiscard]] bool is_float() const;
+  [[nodiscard]] int get_ivalue() const override;
+  [[nodiscard]] float get_fvalue() const override;
 
  private:
   bool _is_float = false;
@@ -153,46 +94,24 @@ class ASTNumberLiteral final : public ASTNode {
 
 class ASTPrefix : public ASTNode {
  public:
-  ASTPrefix() : ASTNode(ASTType::EOF_, 1, op_precedence[ASTType::EOF_], 0) {}
-
-  [[nodiscard]] ASTNode *nud(Parser *parser) override {
-      auto *n = parser->next_expression(_lbp);
-      if (!n) {
-          // TODO: report error
-      } else {
-          _children.emplace_back(n);
-      }
-      return this;
-  }
+  ASTPrefix();
+  [[nodiscard]] ASTNode *nud(Parser *parser) override;
 };
 
 class ASTBinaryNot final : public ASTPrefix {
  public:
-  ASTBinaryNot() : ASTPrefix() {
-      _op = ASTType::BNOT;
-      _lbp = op_precedence[_op];
-  }
+  ASTBinaryNot();
 };
 
 class ASTLogicalNot final : public ASTPrefix {
  public:
-  ASTLogicalNot() : ASTPrefix() {
-      _op = ASTType::LNOT;
-      _lbp = op_precedence[_op];
-  }
+  ASTLogicalNot();
 };
 
 class ASTStringLiteral final : public ASTNode {
  public:
-  explicit ASTStringLiteral(std::string str) : ASTNode(ASTType::STRING_LITERAL, 1,
-                                                       op_precedence[ASTType::STRING_LITERAL],
-                                                       0) {
-      _svalue = std::move(str);
-  }
-
-  [[nodiscard]] std::string get_svalue() const override {
-      return _svalue;
-  }
+  explicit ASTStringLiteral(std::string str);
+  [[nodiscard]] std::string get_svalue() const override;
 
  private:
   std::string _svalue;
@@ -200,34 +119,22 @@ class ASTStringLiteral final : public ASTNode {
 
 class ASTSum final : public ASTInfixBinaryOp {
  public:
-  ASTSum() : ASTInfixBinaryOp() {
-      _op = ASTType::SUM;
-      _lbp = op_precedence[ASTType::SUM];
-  };
+  ASTSum();
 };
 
 class ASTSubtract final : public ASTInfixBinaryOp {
  public:
-  ASTSubtract() : ASTInfixBinaryOp() {
-      _op = ASTType::SUBTRACT;
-      _lbp = op_precedence[ASTType::SUBTRACT];
-  };
+  ASTSubtract();
 };
 
 class ASTMultiply final : public ASTInfixBinaryOp {
  public:
-  ASTMultiply() : ASTInfixBinaryOp() {
-      _op = ASTType::MULTIPLY;
-      _lbp = op_precedence[ASTType::MULTIPLY];
-  };
+  ASTMultiply();
 };
 
 class ASTDivide final : public ASTInfixBinaryOp {
  public:
-  ASTDivide() : ASTInfixBinaryOp() {
-      _op = ASTType::DIVIDE;
-      _lbp = op_precedence[ASTType::DIVIDE];
-  }
+  ASTDivide();
 };
 
 }
