+++ src/base/macro.h
@@ -3,7 +3,9 @@
 
 #include "config.h"
 
-#define _STR_(s) #s
+/// \brief for suppressing unused parameters warning in a function
+#define UNUSED(x) (void)(x)
+
 #ifdef __has_builtin
 #    define TAN_HAVE_BUILTIN(x) __has_builtin(x)
 #else
@@ -12,12 +14,11 @@
 
 // UNLIKELY and LIKELY for optimization
 #if TAN_HAVE_BUILTIN(__builtin_expect) || (defined(__GNUC__) && !defined(__clang__))
-#    define TAN_LIKELY(x) (__builtin_expect(x, 0))
-#    define TAN_UNLIKELY(x) (__builtin_expect(!!(x), 1))
+#   define TAN_LIKELY(x) (__builtin_expect(x, 0))
+#   define TAN_UNLIKELY(x) (__builtin_expect(!!(x), 1))
 #else
-#    define TAN_LIKELY(x) (x)
-#    define TAN_UNLIKELY(x) (x)
+#   define TAN_LIKELY(x) (x)
+#   define TAN_UNLIKELY(x) (x)
 #endif
-#define CONCAT3(x, y, z) x##y##z
-#define CONCAT2(x, y) x##y
+
 #endif // __TAN_SRC_BASE_MACRO_H__
